fixed error bugs

+++ server/models/comments/comments.models.ts
@@ -1,33 +1,37 @@
 import commentEntity from "../comments/comments.entity";
 import {commentType} from "../../types/comment.types"
+import { AppError } from "../../utils/errorHander";
+import { Types } from "mongoose";
 
 class CommentModels{
     async insertComment(comment:commentType){
-        try {
-            return await new commentEntity(comment).save();
-        } catch (error) {
-            throw new Error(error)
-        }
+          return await new commentEntity(comment).save();
     }
 
     async getAllCommentsOfAVideo(video_id:string){
-        try {
-            const result=await commentEntity.find({v_id:video_id});
-            if (!result) {
-                throw new Error("No comments Found");
-            }
-            
-        } catch (error) {
-            throw new Error(error)
+        if (!Types.ObjectId.isValid(video_id)) {
+            throw new AppError("Bad request id is not valid",400);
         }
+        const result=await commentEntity.find({v_id:video_id},{v_id:0,u_id:0});
+        if (!result.length) {
+            throw new AppError("No comments Found",400);
+        }
+        return result;
     }
 
     async deleteComment(commentId:string){
-        try {
-            await commentEntity.findByIdAndDelete(commentId);
-        } catch (error) {
-            throw new Error(error)
+        if (!Types.ObjectId.isValid(commentId)) {
+            throw new AppError("Bad request id is not valid",400);
         }
+        const result=await commentEntity.findByIdAndDelete(commentId);
+        return result;
+    }
+    async getAllComments(u_id:string){
+        if (!Types.ObjectId.isValid(u_id)) {
+            throw new AppError("Bad request id is not valid",400);
+        }
+        const result=await commentEntity.find({u_id:u_id});
+        return result;
     }
 
 }
